Move ChildFactories into classes which use them. Make CommonFilesMetaDataBuilder to handle queries/ list of model building / deduping files. Simplifies CommonFilesMetaData by separating business logic from model and CommonFilesPanel. Removes yet another query by re-using the dataSourceIdMap from CommonFilesPanel for model creation, so now CommonFiles is down to 2 queries total, including UI building.
